+++ source/histograms_2d.cpp
@@ -36,8 +36,8 @@ int main(int argc, char **argv) {
         tree->Add(input_file.c_str());
     }
 
-    const int first = vm["first"].as<int>();
-    const int last =
+    const long long first = vm["first"].as<int>();
+    const long long last =
         vm["last"].as<int>() == 0 ? tree->GetEntries() : vm["last"].as<int>();
 
     if (first > last) {
@@ -68,7 +68,7 @@ int main(int argc, char **argv) {
                      coincidence_matrices[n_matrix].y_axis.maximum));
     }
 
-    for (int i = first; i <= last; ++i) {
+    for (long long i = first; i <= last; ++i) {
         progress_printer(i - first);
 
         tree->GetEntry(i);
